+++ lib/manageiq/providers/ibm_power_hmc.rb
